Move `formatter<std::optional>` into `optional.hpp`
