+++ libtenzir/include/tenzir/optional.hpp
@@ -8,6 +8,8 @@
 
 #pragma once
 
+#include <fmt/format.h>
+
 #include <optional>
 
 namespace tenzir {
@@ -20,3 +22,21 @@ std::optional<T> to_optional(const T* ptr) {
 }
 
 } // namespace tenzir
+
+namespace fmt {
+template <class T>
+struct formatter<std::optional<T>> {
+  template <class ParseContext>
+  constexpr auto parse(ParseContext& ctx) {
+    return ctx.begin();
+  }
+
+  template <class FormatContext>
+  auto format(const std::optional<T>& value, FormatContext& ctx) const {
+    if (!value) {
+      return fmt::format_to(ctx.out(), "none");
+    }
+    return fmt::format_to(ctx.out(), "{}", *value);
+  }
+};
+} // namespace fmt
